Add module NamespaceConvertTable for convert table
